Multiple Ascending Dose Data Set (Duplicates Removed)

Model generated PK and PD data to mimic an orally administered small molecule with various endpoints from continuous to ordinal response and count data. Simulated multiple dose administration ranging from 100 mg to 1600 mg, once per day.

mad_missing_duplicates

Format

A data frame with the following 19 columns:

column 1:	ID	numeric; unique subject ID

column 2:	TIME	numeric; time relative to first drug administration

column 3:	NOMTIME	numeric; nominal time

column 4:	TIMEUNIT	character; unit of TIME

column 5:	AMT	numeric; dosing amount (for dosing events) in mg

column 6:	LIDV	numeric; observation on a linear scale (observation type determined by CMT), units determined by EVENTU column

column 7:	MDV	numeric; missing dependent variable

column 8:	CMT	integer; compartment number (determines observation type):

CMT 1 = Dosing event

CMT 2 = PK concentration

CMT 3 = Continuous response data

CMT 4 = Count response data

CMT 5 = Ordinal response data

CMT 6 = Binary response data

column 9:	NAME	character; description of event

column 10:	EVENTU	character; unit for observation

column 11:	CENS	integer; censored values (0 = not censored, 1 = censored)

column 12:	EVID	integer; event ID (0 = observation, 1 = dosing event)

column 13:	WEIGHTB	numeric; baseline body weight (kg)

column 14:	SEX	character; sex

column 15:	TRTACT	factor; treatment group label

column 16:	DOSE	numeric; randomized dose in mg

column 17:	PROFDAY	numeric; day of profile

column 18:	PROFTIME	numeric; time within PROFDAY

column 19:	CYCLE	numeric; count of drug administrations received