Prediction data frame for polr

Get predictions with standard errors into data frame for use with geom_smooth

# S3 method for polr
predictdf(
  model,
  xseq,
  se,
  level,
  data,
  method,
  formula,
  method.args,
  weight,
  n_boot = 200
)

Arguments

model
object returned from polr

xseq
sequence of x values for which to compute the smooth

se
if TRUE then confidence intervals are returned

level
confidence level for confidence intervals

data
data to fit

method
only works for method MASS::polr

formula
formula to fit

method.args
arguments to pass to method

weight
weights to use for method

n_boot
number of bootstraps to perform for confidence interval calculation, default is 200

Details

predictdf.polr is used by xgx_geom_smooth when method = "polr" to calculate confidence intervals via bootstraps.