xgx_annotate_status {xgxr}	R Documentation

Create a status (e.g. DRAFT) annotation layer

Description

xgx_annotate_status adds a status (e.g. DRAFT) annotation layer to a plot. The text of the annotation can be customized, the default is "DRAFT". The color, location, size, fontface, transparency of the annotation can also be customized.

Usage

xgx_annotate_status(status = "DRAFT", x = Inf, y = Inf,
  color = "grey", hjust = 1.2, vjust = 1.2, fontsize = 7,
  fontface = "bold", alpha = 0.5, ...)

Arguments

status
the text to

x
x location, default Inf (right most point)

y
y location, default Inf (up most point)

color
default "grey"

hjust
horizontal justification, default 1.2

vjust
vertical justification, default 1.2

fontsize
font size to use, default 7

fontface
font style to use, default "bold"

alpha
transparency, default is 0.5

...
other arguments passed on to layer

Value

ggplot layer

Examples

library(ggplot2)  
data         = data.frame(x=1:1000,y=rnorm(1000))
ggplot(data=data,aes(x=x,y=y)) + 
  geom_point() +
  xgx_annotate_status("DRAFT")

[Package xgxr version 0.0.1.005 Index]