xgx_breaks_time {xgxr}	R Documentation

Sets the default breaks for a time axis

Description

xgx_breaks_time sets the default breaks for a time axis, given the units of the data and the units of the plot. It is inspired by scales::extended_breaks

Usage

xgx_breaks_time(data_range, units_plot)

Arguments

data_range
range of the data

units_plot
units to use in the plot

Details

for the extended breaks function, weights is a set of 4 weights for

simplicity - how early in the Q order are you

coverage - labelings that don't extend outside the data: range(data) / range(labels)

density (previously granularity) - how close to the number of ticks do you get (default is 5)

legibility - has to do with fontsize and formatting to prevent label overlap

Value

numeric vector of breaks

References

Talbot, Justin, Sharon Lin, and Pat Hanrahan. "An extension of Wilkinson’s algorithm for positioning tick labels on axes." IEEE Transactions on visualization and computer graphics 16.6 (2010): 1036-1043.

Examples

xgx_breaks_time(c(0, 5), "h")
xgx_breaks_time(c(0, 6), "h")
xgx_breaks_time(c(-3, 5), "h")
xgx_breaks_time(c(0, 24), "h")
xgx_breaks_time(c(0, 12), "h")
xgx_breaks_time(c(1, 4), "d")
xgx_breaks_time(c(1, 12), "d")
xgx_breaks_time(c(1, 14), "d")
xgx_breaks_time(c(1, 50), "d")
xgx_breaks_time(c(1000, 3000), "d")
xgx_breaks_time(c(-21, 100), "d")
xgx_breaks_time(c(-1, 10), "w")

[Package xgxr version 1.0.2 Index]