xgx_check_data {xgxr}	R Documentation

Check data for various issues

Description

xgx_check_data performs a series of checks on a PK or PKPD dataset It was inspired by the dataset preparation table from IntiQuan.

Usage

xgx_check_data(data, covariates = NULL)

Arguments

data,
the dataset to check. Must contain the above columns

covariates,
the column names of covariates, to explore

Details

The dataset must have the following columns

ID = unique subject identifier. USUBJID is another option if ID is not there

EVID = event ID: 1 for dose, 0 otherwise

AMT = value of the dose

TIME = time of the measurement

DV = dependent value (linear scale). will check if LIDV or LNDV are also there if DV is not

YTYPE = data measurement for LIDV. will check if CMT is there, if YTYPE is not

The dataset may also have additional columns

CENS = flag for censoring of the data because it's below the limit of quantification (BLOQ)

MDV = missing dependent variable - will be counted and then filtered out from the data check

Value

data.frame

Examples

covariates <- c("WEIGHTB", "SEX")
check <- xgx_check_data(mad_missing_duplicates, covariates)

[Package xgxr version 1.0.2 Index]