xgx_dirs2char {xgxr}	R Documentation

Append filenames to bottom of the plot

Description

xgx_dirs2char returns a character variable based on the dirs list. The caption gives the filename

Usage

xgx_dirs2char(dirs, include_time = TRUE)

Arguments

dirs
list containing directories and filenames. It must contain five fields

parent_dir = Parent directory containing the Rscript and the Results folder

rscript_dir = Subdirectory ofparent_dir that contains the Rscript used to generate the figure

rscript_name= Name of the Rscript used to generate the figure

results_dir = Subdirectory ofparent_dir where the figure is stored

filename = Filename

#'

include_time
is logical with default TRUE. If TRUE, it includes date/time in the output character

Value

character

Examples

#' 
library(ggplot2)  
dirs = list(Parent_dir   = "/your/parent/path/",
            rscript_dir  = "./Rscripts/",
            rscript_name = "Example.R",
            results_dir  = "./Results/",
            filename     = "your_file_name.png")
caption = xgx_dirs2caption(dirs)

[Package xgxr version 0.0.1.005 Index]