xgx_geom_ci {xgxr}	R Documentation

Plot data with mean and confidence intervals

Description

Plot data with mean and confidence intervals

Usage

xgx_geom_ci(percentile = 0.95, ...)

Arguments

percentile
is the percentile for the confidence interval. default is .95 should be between 0.5 and 1

...
other arguments passed on to layer. These are often aesthetics, used to set an aesthetic to a fixed value, like color = "red" or size = 3. They may also be parameters to the paired geom/stat.

Value

ggplot2 plot layer

Examples

library(ggplot2)  
data = data.frame(x = rep(c(1,2,3),each=20),
y = rep(c(1,2,3),each=20) + rnorm(60))
ggplot(data,aes(x=x,y=y)) + 
 xgx_geom_ci(percentile = .95)

[Package xgxr version 0.0.1.005 Index]