xgx_save_table {xgxr}	R Documentation

Saving table as an image, also labeling the program that created the table and where the table is stored

Description

Saving table as an image, also labeling the program that created the table and where the table is stored

Usage

xgx_save_table(data, dirs = NULL, filename_main = NULL)

Arguments

data
data.frame or table of results

dirs
list of directories. If NULL or if directories missing, there is default behavior below

parent_dir = Parent directory containing the Rscript and the Results folder, default getwd()

rscript_dir = Subdirectory of parent_dir that contains the Rscript used to generate the figure, default "./"

rscript_name= Name of the Rscript used to generate the figure, default "Name_Of_Script_Here.R"

results_dir = Subdirectory ofparent_dir where the figure is stored, default "./"

filename_prefix = prefix of filename to be appended to filename_main

filename_main
main part of the filename, excluding prefix and extension. no default

Value

ggplot2 plot object

Examples

directory = tempdir()
dirs <- list(parent_dir  = directory,
             rscript_dir = directory,
             rscript_name = "example.R",
             results_dir = directory,
             filename_prefix = "example_")
data <- data.frame(x = c(1, 2), y = c(1, 2))
xgx_save_table(data, dirs = dirs, filename_main = "test")

[Package xgxr version 1.0.2 Index]