xgx_scale_x_log10 {xgxr}	R Documentation

log10 scales the x axis with a "pretty" set of breaks

Description

xgx_scale_x_log10 is similar to scale_x_log10. But it uses what we believe to be a nicer spacing and set of tick marks it can be used the same as scale_x_log10

Usage

xgx_scale_x_log10(breaks = xgx_breaks_log10, minor_breaks = NULL,
  labels = xgx_labels_log10, ...)

Arguments

breaks
major breaks, default is a function defined here

minor_breaks
minor breaks, default is a function defined here

labels
function for setting the labels, defined here

...
other arguments passed to scale_x_log10

Value

ggplot2 compatible scale object

Examples

 
conc <- 10^(seq(-3, 3, by = 0.1))
ec50 <- 1
data <- data.frame(concentration  = conc,
                   bound_receptor = 1 * conc / (conc + ec50))
ggplot2::ggplot(data, ggplot2::aes(x = concentration, y = bound_receptor)) + 
ggplot2::geom_point() + 
  ggplot2::geom_line() + 
  xgx_scale_x_log10() +
  xgx_scale_y_reverselog10()
  

[Package xgxr version 1.0.2 Index]