xgx_scale_x_reverselog10 {xgxr}	R Documentation

Reverse-log transform for the x scale.

Description

xgx_scale_x_reverselog10 is designed to be used with data that approaches 100 A common example is receptor occupancy in drug development. It is used when you want even spacing between 90, 99, 99.9, etc.

Usage

xgx_scale_x_reverselog10(...)

Arguments

...
other parameters passed to scale_x_continuous

Value

ggplot2 compatible scale object

Examples

conc <- 10^(seq(-3, 3, by = 0.1))
ec50 <- 1
data <- data.frame(concentration = conc, 
                   bound_receptor = 1 * conc / (conc + ec50))
ggplot2::ggplot(data, ggplot2::aes(y = concentration, x = bound_receptor)) +
  ggplot2::geom_point() +
  ggplot2::geom_line() +
  xgx_scale_y_log10() +
  xgx_scale_x_reverselog10()
 

[Package xgxr version 1.0.2 Index]