xgx_scale_x_time_units {xgxr}	R Documentation

Convert time units for plotting

Description

xgx_scale_x_time_units converts x axis scale from one time unit to another. Supported units include hours, days, weeks, months, and years, which can also be called using just the first letter (h, d, w, m, y).

Usage

xgx_scale_x_time_units(units_dataset, units_plot = NULL,
  breaks = breaks_function, labels = labels_function, ...)

Arguments

units_dataset
units of the input dataset, must be specified by user as "h", "d", "w", "m", or "y"

units_plot
units of the plot, will be units of the dataset if empty

breaks
One of:

NULL for no breaks

waiver() for the default breaks computed by the transformation object

A numeric vector of positions

A function that takes the limits as input and returns breaks as output

labels
One of:

NULL for no labels

waiver() for the default labels computed by the transformation object

A character vector giving labels (must be same length as breaks)

A function that takes the breaks as input and returns labels as output

...
other parameters for ggplot2::scale_x_continuous

xgx_scale_x_time_units

Examples

data         = data.frame(x=1:1000,y=rnorm(1000))
ggplot(data=data,aes(x=x,y=y)) + 
  geom_point() + 
  xgx_scale_x_time_units(units_dataset="hours",units_plot="weeks")

[Package xgxr version 0.0.1.005 Index]