xgx_scale_y_log10 {xgxr}	R Documentation

log10 scales the y axis with a "pretty" set of breaks

Description

xgx_scale_y_log10 is similar to ggplot2::scale_y_log10. But it uses what we believe to be a nicer spacing and set of tick marks it can be used the same as scale_y_log10

Usage

xgx_scale_y_log10(breaks = xgx_breaks_log10,
  minor_breaks = xgx_log_breaks_minor, labels = xgx_labels_log10, ...)

Arguments

breaks
major breaks, default is a function defined here

minor_breaks
minor breaks, default is a function defined here

labels
function for setting the labels, defined here

...
other arguments passed to ggplot2::scale_y_log10

[Package xgxr version 0.0.1.005 Index]