Edit a Rmd Template from xgx

edit_rmd_template_str returns a path to the altered Rmd template

edit_rmd_template_str(
  rmd_str = NULL,
  mapping = NULL,
  rmd_output_path = NULL,
  data_path = NULL,
  multiple_dosing = FALSE,
  pk_cmt = NULL,
  pd_cmt = NULL,
  dose_cmt = NULL,
  steady_state_day = NULL,
  time_between_doses = NULL,
  author_name = NULL,
  add_datetime = TRUE,
  show_explanation = TRUE
)

Arguments

rmd_str

A character string containing the Rmd template raw characters

mapping

A list of column name mappings from the original (template) dataset column names to the corresponding columns in the new dataset

rmd_output_path

A custom output path for the generated Rmd file (This is typically left as `NULL` in order to maintain the hierarchical directory structure of `xgx_autoexplore_output`))

data_path

Path (as a string) to the dataset that is to be analyzed

multiple_dosing

if FALSE use single ascending dose template, if TRUE use multiple

pk_cmt

An integer denoting the "compartment" containing the PK data. The "CMT" column will typically have these integers, where each row may contain either PK or PD data, potentially of different types (continuous, ordinal, etc.)

pd_cmt

An integer denoting the "compartment" containing the PD data, of the desired type (continuous, ordinal, etc.). The "CMT" column will typically have these integers, where each row may contain either PK or PD data

dose_cmt

CMT associated with dosing event

steady_state_day

For multiple ascending dose, what day is steady state rich profile?

time_between_doses

time interval between doses

author_name

The name of the author to be displayed on the template

add_datetime

Boolean indicating additon of a date stamp to the beginnning of the Rmd file

show_explanation

Boolean indicating if the additional explanations (text in between figures) are needed for the user.

Value

A string of the new R markdown template