Check data for various issues

xgx_check_data performs a series of checks on a PK or PKPD dataset It was inspired by the dataset preparation table from IntiQuan.

xgx_check_data(data, covariates = NULL)

Arguments

data,

the dataset to check. Must contain the above columns

covariates,

the column names of covariates, to explore

Value

data.frame

Details

The dataset must have the following columns

ID = unique subject identifier. USUBJID is another option if ID is not there

EVID = event ID: 1 for dose, 0 otherwise

AMT = value of the dose

TIME = time of the measurement

DV = dependent value (linear scale). will check if LIDV or LNDV are also there if DV is not

YTYPE = data measurement for LIDV. will check if CMT is there, if YTYPE is not

The dataset may also have additional columns

CENS = flag for censoring of the data because it's below the limit of quantification (BLOQ)

MDV = missing dependent variable - will be counted and then filtered out from the data check

Examples

covariates <- c("WEIGHTB", "SEX")
check <- xgx_check_data(mad_missing_duplicates, covariates)
#> Warning: Setting YTYPE column equal to CMT
#> Warning: Setting DV column equal to LIDV
#> removing  points with MDV==1 & EVID==0 from dataset
#> 
#> DATA SUMMARY
#> 
#> ------------------------------------------------------------------------------------------------------------
#> Category      Description                                          YTYPE   Statistic                        
#> ------------- ---------------------------------------------------- ------- ---------------------------------
#> Patients      Number of Patients                                   -       60                               
#> 
#> MDV           Number of patients with zero PK or PD observations   all     0                                
#> 
#> MDV           Number of Missing Data Points (MDV==1 and EVID==0)   1       0                                
#> 
#> MDV           Number of Missing Data Points (MDV==1 and EVID==0)   2       99                               
#> 
#> MDV           Number of Missing Data Points (MDV==1 and EVID==0)   3       0                                
#> 
#> MDV           Number of Missing Data Points (MDV==1 and EVID==0)   4       0                                
#> 
#> MDV           Number of Missing Data Points (MDV==1 and EVID==0)   5       0                                
#> 
#> MDV           Number of Missing Data Points (MDV==1 and EVID==0)   6       0                                
#> 
#> Dose          Number of non-zero doses                             -       300                              
#> 
#> Dose          Number of zero doses (AMT==0)                        -       60                               
#> 
#> Dose          Number of patients that never received drug          -       10                               
#> 
#> DV            Number of Data Points                                1       360                              
#> 
#> DV            Number of Data Points                                2       1309                             
#> 
#> DV            Number of Data Points                                3       600                              
#> 
#> DV            Number of Data Points                                4       600                              
#> 
#> DV            Number of Data Points                                5       600                              
#> 
#> DV            Number of Data Points                                6       600                              
#> 
#> DV            Number of Data Points per Individual                 1       min = 6, median = 6, max = 6     
#> 
#> DV            Number of Data Points per Individual                 2       min = 26, median = 26, max = 27  
#> 
#> DV            Number of Data Points per Individual                 3       min = 10, median = 10, max = 10  
#> 
#> DV            Number of Data Points per Individual                 4       min = 10, median = 10, max = 10  
#> 
#> DV            Number of Data Points per Individual                 5       min = 10, median = 10, max = 10  
#> 
#> DV            Number of Data Points per Individual                 6       min = 10, median = 10, max = 10  
#> 
#> DV            Number of Data Points with zero value (DV==0)        1       0                                
#> 
#> DV            Number of Data Points with zero value (DV==0)        2       0                                
#> 
#> DV            Number of Data Points with zero value (DV==0)        3       0                                
#> 
#> DV            Number of Data Points with zero value (DV==0)        4       75                               
#> 
#> DV            Number of Data Points with zero value (DV==0)        5       0                                
#> 
#> DV            Number of Data Points with zero value (DV==0)        6       423                              
#> 
#> DV            Number of Data Points with NA (is.na(DV))            1       0                                
#> 
#> DV            Number of Data Points with NA (is.na(DV))            2       8                                
#> 
#> DV            Number of Data Points with NA (is.na(DV))            3       0                                
#> 
#> DV            Number of Data Points with NA (is.na(DV))            4       0                                
#> 
#> DV            Number of Data Points with NA (is.na(DV))            5       0                                
#> 
#> DV            Number of Data Points with NA (is.na(DV))            6       0                                
#> 
#> DV+TIME       Multiple measurements at same time                   1       0                                
#> 
#> DV+TIME       Multiple measurements at same time                   2       32                               
#> 
#> DV+TIME       Multiple measurements at same time                   3       0                                
#> 
#> DV+TIME       Multiple measurements at same time                   4       0                                
#> 
#> DV+TIME       Multiple measurements at same time                   5       0                                
#> 
#> DV+TIME       Multiple measurements at same time                   6       0                                
#> 
#> CENS          Number of Censored Data Points                       1       0 (0%)                           
#> 
#> CENS          Number of Censored Data Points                       2       9 (1%)                           
#> 
#> CENS          Number of Censored Data Points                       3       4 (1%)                           
#> 
#> CENS          Number of Censored Data Points                       4       4 (1%)                           
#> 
#> CENS          Number of Censored Data Points                       5       4 (1%)                           
#> 
#> CENS          Number of Censored Data Points                       6       4 (1%)                           
#> 
#> All Columns   Negative Values (number)                             -       DV:12                            
#> 
#> All Columns   Missing Values (number)                              -       LIDV:368, DV:368                 
#> ------------------------------------------------------------------------------------------------------------
#> 
#> CONTINUOUS COVARIATES
#> 
#> -----------------------------------------------------------
#> Covariate   Nmissing   min    25th   median   75th    max  
#> ----------- ---------- ------ ------ -------- ------- -----
#> WEIGHTB     0          52.8   69.2   78.9     89.85   109  
#> -----------------------------------------------------------
#> 
#> CATEGORICAL COVARIATES
#> 
#> -----------------------------------------------------------
#> Covariate   Nmissing   Ndistinct   Value (Count)           
#> ----------- ---------- ----------- ------------------------
#> SEX         0          2           Female (30), Male (30)  
#> -----------------------------------------------------------
#> 
#> POSSIBLE DATA ISSUES - FIRST 6 RECORDS
#> 
#> ----------------------------------------------------
#> Data_Check_Issue   ID   TIME     DV   CENS   YTYPE  
#> ------------------ ---- -------- ---- ------ -------
#> DV == 0            1    -0.005   0    0      6      
#> 
#> DV == 0            1    24.33    0    0      6      
#> 
#> DV == 0            1    48.37    0    0      6      
#> 
#> DV == 0            1    72.2     0    0      6      
#> 
#> DV == 0            1    96.26    0    0      6      
#> 
#> DV == 0            1    120.2    0    0      6      
#> ----------------------------------------------------
#> 
#> The following columns contained missing values
#> LIDV:368, DV:368