xgx_scale_x_reverselog10 {xgxr}	R Documentation

Reverselog transform for the x scale.

Description

xgx_scale_x_reverselog10 is designed to be used with data that approaches 100 A common example is receptor occupancy in drug development. It is used when you want even spacing between 90, 99, 99.9, etc.

Usage

xgx_scale_x_reverselog10(...)

Arguments

...
other parameters passed to ggplot2::scale_x_continuous

Examples

library(ggplot2)  
conc = 10^(seq(-3,3,by = 0.1))
EC50 = 1
data = data.frame(concentration  = conc, bound_receptor = 1*conc/(conc+EC50))
ggplot(data,aes(y=concentration,x=bound_receptor)) + 
geom_point() + 
 geom_line() + 
 xgx_scale_y_log10() +
 xgx_scale_x_reverselog10()

[Package xgxr version 0.0.1.005 Index]